chore: update to latest versions of broadway, ex_aws_dynamo, and telemetry
Broadway v1.0 has no breaking changes with v0.6.2 according to the changelog, so this seems alright.
ex_aws_dynamo v4 is a breaking change over v3 when reading and writing empty strings as values. Previously empty strings were treated as nil values, so a write of an empty string would come back as nil. v4 makes it so that a write of an empty string will come back as an empty string on read - but previosly written empty strings treated as nil will still come back as nil until overwritten. This does not seem to be a problem for kcl_ex, as I don't think it stores empty strings.
Telemetry v1.0 marked the library as stable. No breaking changes.