feat(pla-286): add can_acquire_lease callback
This callback is checked before attempting to take or renew a lease. If it returns true (the default), we proceed as normal. If it returns false, then no lease acquisition or renewal is attempted, and any running pipelines are halted.

We will use this hook to determine if the currently running pod ought to be running the given shard, the idea being that we can convert the list of running pods into a hash ring, hash the shard id, and determine which pod ought to have the lease.

In production, we will cache the k8s replicaset information so that this lease acquisition check is very fast, and put the whole thing behind a feature flag so we can roll this out incrementally.