fix: retry dynamodb checkpoints (#19)
Introduce a loop (with jittery exponential backoff retries) of up to 5 retries when provisioned write throughput is exceeded on dynamo checkpoint table writes. If after 105attempts we continue to fail, then raise an exception and let it crash.

Note that as-is, kcl_ex does not handle producer crashes properly, as the shard producer will fail to restart. A more OTP-friendly approach with monitors is warranted here for robustness, but for the sake of fixing this edge we are using this retry loop in favor of a complete refactor.